PyFiles/requirements.txt
torch==2.1.0
numpy==1.24.3
torchvision==0.16.0
Pillow==9.4.0
python==3.8.10